requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
facenet-pytorch
numpy
pandas
kagglehub
matplotlib
scikit-learn
opencv-python
torch
torchvision
torchmetrics
seaborn
accelerate
timm
faiss-cpu
